requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
authlib==1.3.0
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
itsdangerous==2.2.0
cryptography==42.0.7
email-validator==2.1.1
